ci(react-native): separate out some common functions from fixture generation script
